Make OptionEditor available to everyone
